Make icw unicode friendly
Updated to decode input using utf-8-sig, which is apparently a UTF8
version that Microsoft made with some funky characters at the beginning
of the file. It seems like this was probably causing a rare error for
users that edited their file in Notepad on a PC. Also now uses
unicode_literals, debug statements use unicode, the unicode workaround
for older versions of the csv module
(https://docs.python.org/2/library/csv.html#examples), and includes a
unicode test file.