Use chardet module for encoding errors
After all the work to make things unicode happy, I was still seeing
several UnicodeDecodeErrors in my logs from users using `latin1` and
what have you instead of `utf8`. I decided to give a shot to a new
strategy: since `chardet` is pure python, it can be uploaded to GAE as
part of the package. However, to keep things as speedy as possible, I
decided to try EAFP and assume utf8 unless I get a UnicodeDecodeError,
in which case change encoding to whatever chardet thinks. One downside
to this strategy is that the unicode_csv_reader (see b8f45f5) generates
output line by line, meaning that it could potentially read the first
several lines of a file in the utf8 default before it has an exception
and uses chardet to figure out the *real* encoding. I'm not sure how
often that would actually cause a problem, so I'm going to give this a
shot.